better component design
